xzceb-flask_eng_fr-master/final_project/requirements.txt
Flask
Flask-WTF
ibm-watson
python-dotenv
pyJWT
pylint
